+++ synthesizer/src/snark/certificate/bytes.rs
@@ -20,7 +20,7 @@ impl<N: Network> FromBytes for Certificate<N> {
     /// Reads the certificate from a buffer.
     fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
         // Read the version.
-        let version = u16::read_le(&mut reader)?;
+        let version = u8::read_le(&mut reader)?;
         // Ensure the version is valid.
         if version != 0 {
             return Err(error("Invalid certificate version"));
@@ -36,7 +36,7 @@ impl<N: Network> ToBytes for Certificate<N> {
     /// Writes the certificate to a buffer.
     fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
         // Write the version.
-        0u16.write_le(&mut writer)?;
+        0u8.write_le(&mut writer)?;
         // Write the bytes.
         self.certificate.write_le(&mut writer)
     }
